+++ filament/include/filament/LightManager.h
@@ -325,6 +325,18 @@ public:
              * Higher values may not be available depending on the underlying hardware.
              */
             uint8_t msaaSamples = 1;
+
+            /**
+             * Standard deviation of the VSM blur. Zero do disable.
+             * The maximum value is 21, which corresponds to a gaussian blur filter width
+             * of 125 pixels. The relation between the filter width and the standard deviation
+             * is roughly: stddev = (kernelWidth + 1) / 6.
+             * Some common values for blurStandardDeviation:
+             *  3x3 gaussian : 0.6667
+             *  5x5 gaussian : 1.0
+             *  9x9 gaussian : 1.6667
+             */
+            float blurStandardDeviation = 0.0f;
         } vsm;
     };